Mocking on UpdateCurrAcc.test (addMovement mock)
