+++ src/updateFunctions/updateFunctions.test.js
@@ -8,100 +8,7 @@ import { accounts, date } from '../data';
 /// 24.02.2021 Working on tests
 
 
-////////// Testing updateCurrentAcc()
-// I'm using here the addMovement() as a subfunction , I'd already tested it...(see below)
-describe("updateCurrAcc()", () => {
-    const currentAcc = {
-        owner: 'Test Acc',
-        balance: 10000,
-        movements: [{
-            amount: -200,
-            date: date,
-            transactionTyp: "casual",
-            sender: 'ta',
-            recepient: 'km',
-            message: `i don't know`
-        }],
-        debt: [{ to: "km", value: 1000 }],
-        lended: [],
-    };
-    const accs = { currentAcc }
-    test(`update CurrentAcc by repayment`, () => {
-        const result = updateCurrAcc("km", "repayment", 1000, "nothing", accs, date, false)
-            // Assertion 
 
-        expect(result).toEqual({
-            owner: 'Test Acc',
-            balance: 9000,
-            movements: addMovement(accs.currentAcc.username, "km", -1000, "repayment", "nothing", date, accs.currentAcc),
-            debt: [],
-            lended: [],
-
-        })
-
-    });
-
-    test(`update CurrentAcc by lend when acc exists in lended array`, () => {
-        const currentAcc = {
-            owner: 'Test Acc',
-            balance: 9000,
-            movements: [{
-                amount: -200,
-                date: date,
-                transactionTyp: "casual",
-                sender: 'ta',
-                recepient: 'km',
-                message: `i don't know`
-            }],
-            debt: [],
-            lended: [{ to: "km", value: 1000 }],
-        };
-        const accs = { currentAcc }
-
-        const result = updateCurrAcc("km", "lend", -1000, "nothing", accs, date, true)
-
-        // Assertion 
-
-        expect(result).toEqual({
-            owner: 'Test Acc',
-            balance: 8000,
-            movements: addMovement(accs.currentAcc.username, "km", -1000, "lend", "nothing", date, accs.currentAcc),
-            debt: [],
-            lended: [{ to: "km", value: 2000 }],
-
-        })
-
-    });
-
-    test(`update CurrentAcc by lend when acc doesn't exist in lended array`, () => {
-        const currentAcc = {
-            owner: 'Test Acc',
-            balance: 9000,
-            movements: [{
-                amount: -200,
-                date: date,
-                transactionTyp: "casual",
-                sender: 'ta',
-                recepient: 'km',
-                message: `i don't know`
-            }],
-            debt: [],
-            lended: [],
-        };
-        const accs = { currentAcc }
-        const result = updateCurrAcc("km", "lend", -1000, "nothing", accs, date, false)
-            // Assertion 
-
-        expect(result).toEqual({
-            owner: 'Test Acc',
-            balance: 8000,
-            movements: addMovement(accs.currentAcc.username, "km", -1000, "lend", "nothing", date, accs.currentAcc),
-            debt: [],
-            lended: [{ to: "km", value: 1000 }],
-
-        })
-    })
-});
 
 /// Testing findDeleteOrDecrease()
 describe("findDeleteOrDecrease()", () => {
